requirements.txt
gunicorn
flask
psycopg2
bcrypt
flask-mail
pillow
qrcode
boto3